Reconnect watcher when stream closes or server shuts down (#2)
There're two cases when watcher stream could be closed:

- When the cached enabled for Auth Server. It initiates stream closing
  every 600 seconds. We check this case using IsEOF(err).
- When Auth Server goes offline. Then we should start another watcher.
  We check this case using IsConnectionProblem(err).

Further, we should consider doing GetRequests(PENDING) after every
successful reconnection not to miss any requests created between
disconnection and reconnection. Not sure how to do this using the
current API though. Must we fetch all the pending requests and match
them against their presence in the cache? How we can be sure that
there's not so large amount of records?