Consider graceful shutdown for outgoing HTTP-requests too.
We already have `httpServer.Shutdown()` to ensure that all incoming
HTTP requests are finished before we terminate the process.

However, we don't do the same for outgoing HTTP requests. On
termination, all the background operations are simply halted by
context cancellation. To challenge this problem, I extended the use of
existing `WaitGroup` and introduced the `spawn(func(ctx)...)` method
to execute goroutines as jobs that could be waited to finish.