jirabot: Migrate to using issue property instead of custom field (#55)
Custom fields create many inconveniences. They're visible in the issue
view and in fact are editable by user. To work, custom field should also
be explicitely activated in the issue type settings which is cumbersome.

There's in fact a better tool for a job - JIRA entity properties which
allows to store arbitrary JSON blob inside JIRA objects.

However, go-jira library lacks the support of issue properties so I had
to temporarily patch the package. Hope we'll get rid of it soon.

Closes #55