Remove approve/deny buttons from Slack plugin completely (!!!).
This was a @klizhentas idea to make Slack plugin notification only.

Every slack message includes either a link to the request on Web UI
(only in Enterprise version) or `tctl` command snippet for opensoucre version.

Recepients of the notifications are either set statically in the config
file (opensource version) or determined dynamically from `SuggestedReviewers`.