mattermost: Rewrite plugin. Remove approve/deny buttons completely (!!!)
This was a @klizhentas idea to make Mattermost plugin notification only.

Every message includes either a link to the request on Web UI
(only in Enterprise version) or `tctl` command snippet for opensoucre version.

Recepients of the notifications are either set statically in the config
file (opensource version) or determined dynamically from `SuggestedReviewers`.