Fix sed command for BSD version
The BSD version of the sed command is slightly different from the
GNU version, so let's decide at the beginning which format to use
based on the fact that BSD sed doesn't have a --version option.

This might sound a bit hacky, but it's actually more stable than
testing for `uname -s` (which will only work for the systems we
test for).

Tested with GNU sed on Linux and BSD sed on MacOS.