Access/Email: add StartTLS Policy to SMTP config
When sending an email over SMTP we can define how we communicate to the server.
It starts with plain text, but we can then upgrade the connection to using TLS,
before sending the actual email data

We can use:
- disabled/NoStartTLS: the communication will be in plain text between client
and server
- opportunistic: we'll try to use TLS, however if the server refuses to use an
encrypted channel, we fallback to plain text
- mandatory: we'll only send the email using TLS, if the server refuses, we
won't be sending anything

This PR aims to introduce such configuration
Previously we were always using the 'mandatory' option, which will continue to
be the default from now on.

Using the port 465 will override everything we said above and will always start
the connection using TLS
There's no plain text, neither a negotation to upgrade the connection

Fixes #484