terraform: provider fails to create resource with cache on
When creating an app, if the Teleport Auth service has cache enabled,
the provider will fail but the app is created anyways.

This happens because when the provider receives the CreateApp
instruction from terraform, it will:
- check if app exists, if it does: error
- create the app
- get the app state to store it locally (in tf state)

Between the second and third step, we will most likely get a `not found`
error, because the app is not yet visible due to caching
eviction/refresh constraints.

However, given the second step (app creation) was executed succefully,
the app already exists in Teleport

This should happen for other resources as well, however, we are only
adding a test to the app as an example (the fix will be applied to every
resource)