Auto generate TF schema when updating the version
When we release plugins version, we also bump the Teleport API version
to match the current release.

This usually adds new fields to the API
In order for those fields to be used in the Terraform's Teleport
Provider, we must add them to the schema definition
This is accomplished by code generation, through the `gen-tfschema`
script present in the `terraform/Makefile`.

Currently we do this in two phases:
- PR with the version upgrade
- PR with the terraform schema re-generation

This PR aims to reduce this to a single step, by addind the
`gen-tfschema` to the `update-version` command.