access/msteams: Add uninstall command (#637)
Followup of the `access/msteams` PR #617, part of https://github.com/gravitational/teleport/issues/11818

The way the personal teams notification system work is the following:
- add app to MsTeams instance catalog
- install the app without user interaction to all recipients (apps have to be enabled per-user)
- get the canal ID that was created during installation
- speak with the user through this cana

However if we want to uninstall the App (something went wrong, a new app instance was created, ...) we need to disable it for all users that were contacted or in the recipient list. This commit introduces a command that will uninstall the apps for all users in the `recipients` list.