Fix wait for ssh node ready
We were only checking for the following message to be present
`Service <xyz> is starting on <host:port>`

But that's not exactly when the port is open and acepting new clients.
We need to wait a little longer, so we are now waiting for a new line to
be appear:
```
The new service has started successfully.
```

We also added a couple extra checks to ensure the service is ready to
receive clients.