Add Machine ID Bot resource for the terraform provider
This adds a new `teleport_bot` resource to the Terraform provider
which allows creation of Teleport bots.

This resource is a bit unusual as bots do not have any server-side
state of their own. As such, this includes a small hand-written
resource. It doesn't support updates as bot fields are currently
immutable; fields users might want to change are marked with
`tfsdk.RequiresReplace()` causing the resource to be recreated upon
modification.

No data source is currently provided, due to the previously mentioned
lack of a server side resource to query for that returns the data
necessary to populate the schema.

A new usage example is also included for this resource type.