Add Machine ID Bot resource for the terraform provider (#669)
* Add Machine ID Bot resource for the terraform provider

This adds a new `teleport_bot` resource to the Terraform provider
which allows creation of Teleport bots.

This resource is a bit unusual as bots do not have any server-side
state of their own. As such, this includes a small hand-written
resource. It doesn't support updates as bot fields are currently
immutable; fields users might want to change are marked with
`tfsdk.RequiresReplace()` causing the resource to be recreated upon
modification.

No data source is currently provided, due to the previously mentioned
lack of a server side resource to query for that returns the data
necessary to populate the schema.

A new usage example is also included for this resource type.

* Add tests for the Bot resource

This adds basic acceptance tests for the bot resource, fixes a
missing ID field needed by the testing framework, and marks the
`traits` field with `RequiresReplace`.

* Fix some lints, add missing cleanup check for bot user

* Attempting to fix the flaky test

* Remove variable reference apparently causing CI to fail

* Address review feedback

Makes token_id a sensitive and required field, and adds some detail
to some TODO comments.

* Remove TODO

* Push out expiration dates