Terraform: deprecate field `teams_to_logins`
Teleport is moving from `teams_to_logins` to `teams_to_roles`.
As of V10, both are configurable but Terraform was requiring the
deprecated one(`teams_to_logins`).

This commit allows for both to be set and paves the way for the removal
of the property `teams_to_logins` when we release v12.