Terraform: allow identity to be provided as bas64
When configuring the Teleport Provider we have multiple methods.
One of them is to provide the Identity as a file path.

With this PR we can now provide the identity as a base64 encoded string.

This PR also adds a `dial_timeout_duration` configuration for the
provider to ensure we can set our own limit instead of always using 30s.