access/msteams: Fix pluginData encoding and token renewal (#692)
Fixes https://github.com/gravitational/teleport/issues/18068

Encoding/decoding of messages in pluginData was breaking if the data contained `/` or `,`. This was the case in the issue above as the customer used the teams URL channel.

To fix this I switched to a more robust encoding: each message is json-encoded and `messages` is a list of the base64-encoded messages. This will increase the storage size needed for the plugin data by 33%. I don't think this will be an issue.

The commit also contains the following changes:
- better handling for empty/wrong plugin data (fail fast with better errors)
- do not swallow errors when retrying in the `cas`
- Fix token renewal logic (the token was either always renewing, or never renewing)

Co-authored-by: Marco André Dinis <marco.dinis@goteleport.com>