Decrease TestDecorr flakiness
Fixes an off-by-one assertion by letting the sleep duration be as great
as the configured max/cap.

It also explains why we need `runtime.Gosched`.

This test relies on some runtime assumptions that fail sometimes.
I tried to keep the same base code because it seems to be correct, just
that it's hard to prove.