Use quintush/helm-unittest for helm unittests
This plugin is not the latest, most of its feature have been brought
back to mainline helm-unittest recently, but this is the version we use
on the `teleport` charts. We'll need to change the plugin version again
the future (and this wil cause another breaking change :/) but for now
this is better for us to have the same test tools for `teleport` and
`teleport-plugins`.