TF Ref Docs: Update to 13.0.3
We released 13.0.3
This PR updates the reference docs for it.

It also changes the way we escape the fields description.
It was using html escape previsouly, but then some words would trigger
the cspell lint check we have in `Lint(Docs)` CI Step.
Eg `doesn't` became `doesn&#39;t` and `doesn` isn't a valid word.