Pin Golang Docker image to Debian 11
Currently, we're using the latest Docker Go image without specifying a version which defaults to the latest. Debain 12 was releases recently which upgraded our build enviorment.
This PR pins down the OS version, which should solve the issue.

Distroless images are still not upgraded to use Debian 12, hence the downgrade for building images. We should upgrade eveything once this happens.