Pin Golang Docker image to Debian 11 (#854)
Currently, we're using the latest Docker Go image without specifying a version that defaults to the latest. Debian 12 was released recently, which upgraded our build environment. This PR pins down the OS version, which should solve the issue.

Distroless images are still not upgraded to use Debian 12, hence the downgrade for building images. We should upgrade everything once this happens.