Add standard annotation support to Slack plugin
The `teleport-kube-agent` and `teleport-cluster` charts have a standard
pattern of

```
annotations:
  resourceType1:
    keyA: valueA
  resourceType2:
    keyB: valueB
```

etc. The Slack plugin chart currently doesn't support this, having a top
level `podAnnotations` value (undocumented on
https://goteleport.com/docs/reference/helm-reference/teleport-plugin-slack/)
only.

I need to be able to add annotations to the `ConfigMap` and `Deployment`
so that our install of stakater/reloader picks up when the
`roleToRecipients` map changes (because we've added a new role that goes
to a non-default Slack channel) and restarts the container so that it
fetches its new config.

This was a problem we ran into recently, where it took a bit of time to
realise that Slack notifications for some Teleport roles were going to
the wrong channel.

This PR brings the Slack plugin into line with the more recent versions
of the core cluster and agent helm charts, and allows setting
annotations on any of the created resources via a consistent block.

For backwards compatability, the old `podAnnotations` value has been
retained but the newer `annotations.pod` version takes priority.