Add in jwt_claims, update okta_import_rules misnamed field.
The Okta import rules "match" field was misnamed in tests and in the example.
This has been corrected. Additionally, the new `jwt_claims` field was imported
while running `gen-tfschema`, so that has been added to this PR as well.

The tests have been expanded to cover the whole Okta import rules object,
which is why this field misnaming incident in the example wasn't caught in
the first place.