Prepare for Teleport V14
Update Teleport version to 14.0.0 from 13.8.8. The former has now been
released and is available.

Update the Go version to 1.21.1, as used by Teleport 14.

Fix a couple of tests that used `lib/client.Client.CreateAccessRequest()`
to use `CreateAccessRequestV2()` as that supersedes the former. The fix
is as per https://github.com/gravitational/teleport/pull/31760.

Co-authored-by: Cam Hutchison <camh@goteleport.com>