Specify RoleV6...KuberbetesResources.Verbs as computed
Specify `RoleV6.Spec.Allow.KubernetesResources.Verbs` as a computed
field and re-generate the terraform provider. This is needed as the
verbs field has a default value of `[ "*" ]`, which causes the error:

    role_test.go:244: Step 7/8 error: Error running apply: exit status 1

        Error: Provider produced inconsistent result after apply

        When applying changes to teleport_role.upgrade, provider
        "provider[\"registry.terraform.io/hashicorp/teleport\"]" produced an
        unexpected new value: .spec.allow.kubernetes_resources[0].verbs: was null,
        but now cty.ListVal([]cty.Value{cty.StringVal("*")}).

        This is a bug in the provider, which should be reported in the provider's own
        issue tracker.

Ideally this would not be computed but would have a default specified
instead, but it does not look like we can do that with the current
generator.