Prepare for Teleport V14 (#920)
* Prepare for Teleport V14

Update Teleport version to 14.0.0 from 13.8.8. The former has now been
released and is available.

Update the Go version to 1.21.1, as used by Teleport 14.

Fix a couple of tests that used `lib/client.Client.CreateAccessRequest()`
to use `CreateAccessRequestV2()` as that supersedes the former. The fix
is as per https://github.com/gravitational/teleport/pull/31760.

Co-authored-by: Cam Hutchison <camh@goteleport.com>

* Update go.mod to Teleport 14 tag

Update the go.mod file to use Teleport 14 from the tag not the branch by
running:

   make update-teleport-dep-version VERSION=14.0.0

The codeql workflow still seems to use an older version of Go which
fails to parse the new `toolchain` directive in `go.mod`, so add a "Set
up Go" step the same as the Teleport repo to ensure the latest version
is installed for the workflow.

* Update protoc-gen-terraform version to 2.1.0

Use 2.1.0 as this is the current latest of
gravitational/protoc-gen-terraform, and the instructions say to use the
latest.

* Specify RoleV6...KuberbetesResources.Verbs as computed

Specify `RoleV6.Spec.Allow.KubernetesResources.Verbs` as a computed
field and re-generate the terraform provider. This is needed as the
verbs field has a default value of `[ "*" ]`, which causes the error:

    role_test.go:244: Step 7/8 error: Error running apply: exit status 1

        Error: Provider produced inconsistent result after apply

        When applying changes to teleport_role.upgrade, provider
        "provider[\"registry.terraform.io/hashicorp/teleport\"]" produced an
        unexpected new value: .spec.allow.kubernetes_resources[0].verbs: was null,
        but now cty.ListVal([]cty.Value{cty.StringVal("*")}).

        This is a bug in the provider, which should be reported in the provider's own
        issue tracker.

Ideally this would not be computed but would have a default specified
instead, but it does not look like we can do that with the current
generator.

---------

Co-authored-by: Cam Hutchison <camh@goteleport.com>