ci: Build darwin tag builds on linux
Build the tag builds of the terraform provider and event-handler for
darwin on linux. These are pure Go programs so can easily be
cross-compiled using GOOS/GOARCH. The "tricky" part is the universal
binary, solved by using https://github.com/konoui/lipo - a Go version of
lipo (the tool that constructs universal binaries) that can run on
Linux.