ci: Combine linux and darwin tag pipelines
Combine the pipelines for tag builds of linux and darwin builds of the
terraform provider and event-handler. These are pure Go programs so can
easily be cross-compiled using GOOS/GOARCH. The "tricky" part is the
universal binary, solved by using https://github.com/konoui/lipo - a Go
version of lipo (the tool that constructs universal binaries) that can
run on Linux.

This removes Apple Mac machines from the release pipeline of the
plugins.