Update base docker images to use Debian 12 distroless base image (#928)
Instead of chasing ever-changing commit hashes, just use the apppropriate
tag for the distroless image. This aligns things to how `teleport` is handled.

Additionally, standardize on Debian 12 for everything (instead of a mix of 11 and 12).
Also, use `static` over `base`. This means no `glibc` or `libssl`, which should be fine for these plugins.