Terraform Provider: Use updated APIs for Users, Roles, & Auth Connectors
The resources now all expose a Create, Update, Upsert RPC which
returns the resource and an error. The generated code now uses Create
for creating resources and Upsert for updating resources. The Update
RPC is not used because it enforces optimistic locking which breaks
the ownership model expected by the provider. Additionally, the
GetUser RPC now takes a context so the hack to accommodate it has
been removed.