Release 15.3.1 (#1066)
* Release 15.3.1

* access/email: Embed example TOML file

The teleport repo changed to remove the exported `email.ExampleConfig`
variable, but we depend on it here. Add the example config as a file and
embed it in main.

This plugin is on its last days as it gets migrated to the teleport
repo, so duplicating this example config here for a short time will not
become a maintenance issue.